portfolio_projects/breeding-data-pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
pytest>=7.3.0
